change pickup day comment
