+++ config/asset_compress.ini
@@ -73,8 +73,8 @@ files[] = blueimp-file-upload/js/vendor/jquery.ui.widget.js
 files[] = helper.js
 files[] = plugin:Admin:js/upload.js
 files[] = plugin:Admin:js/dialog-customer.js
-files[] = plugin:Admin:js/dialog-order.js
 files[] = plugin:Admin:js/dialog-order-detail.js
 files[] = plugin:Admin:js/dialog-product.js
 files[] = plugin:Admin:js/admin.js
 files[] = plugin:Admin:js/timebased-currency.js
+files[] = plugin:Admin:js/pickup-day.js
