+++ plugins/Admin/src/Locale/de_DE/admin.po
@@ -4,8 +4,8 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
-"POT-Creation-Date: 2018-08-06 16:33+0200\n"
-"PO-Revision-Date: 2018-08-06 15:18+0200\n"
+"POT-Creation-Date: 2018-08-07 09:25+0200\n"
+"PO-Revision-Date: 2018-08-07 09:26+0200\n"
 "Last-Translator: \n"
 "Language-Team: LANGUAGE <EMAIL@ADDRESS>\n"
 "Language: de_DE\n"
@@ -288,6 +288,9 @@ msgstr "Der Preis des bestellten Produktes {0} (Anzahl: {1}) wurde erfolgreich v
 msgid "Ordered_price_adapted"
 msgstr "Preis angepasst"
 
+msgid "The_pickup_day_comment_of_{0}_was_changed:"
+msgstr "Der Abholtag-Kommentar des Mitglieds {0} wurde geändert: "
+
 msgid "Product_{0}_with_a_price_of_{1}_from_{2}_was_successfully_cancelled."
 msgstr "Produkt {0} (Preis: {1}) aus vom {2} wurde erfolgreich storniert."
 
@@ -296,12 +299,6 @@ msgstr "Produkt storniert"
 
 msgid "{0,plural,=1{product_was_cancelled_succesfully.} other{products_were_cancelled_succesfully.}}"
 msgstr "{0,plural,=1{Produkte wurden erfolgreich storniert.} other{Produkte wurden erfolgreich storniert.}}"
-
-msgid "The_comment_of_the_order_number_{0}_by_{1}_was_changed:"
-msgstr "Der Kommentar der Bestellung Nr. {0} von {1} wurde geändert:"
-
-msgid "The_date_of_the_order_{0}_of_{1}_was_changed_to_{2}."
-msgstr "Die Bestellung {0} von {1} wurde auf den {2} rückdatiert."
 
 msgid "Home"
 msgstr "Home"
@@ -1500,8 +1497,8 @@ msgstr "Es wurden keine Bestellungen für den Abholtag {0} gefunden."
 msgid "No_orders_found_for_delivery_period_{0}_-_{1}."
 msgstr "Es wurden keine Bestellungen für den Lieferzeitraum {0} bis {1} gefunden."
 
-msgid "Change_order_status"
-msgstr "Bestellstatus ändern"
+msgid "All_products_picked_up?"
+msgstr "Alle Produkte abgeholt?"
 
 msgid "Back_to_deposit_account"
 msgstr "Zurück zum Pfandkonto"
@@ -1752,6 +1749,15 @@ msgstr "{0,plural,=1{Datensatz} other{Datensätze}}"
 msgid "The_delivered_weight_will_eventually_be_adapted_which_means_the_price_can_change_slightly."
 msgstr "Das tatsächlich gelieferte Gewicht wird evtl. noch angepasst, d. h. der Preis kann sich noch geringfügig ändern."
 
+#~ msgid "The_comment_of_the_order_number_{0}_by_{1}_was_changed:"
+#~ msgstr "Der Kommentar der Bestellung Nr. {0} von {1} wurde geändert:"
+
+#~ msgid "The_date_of_the_order_{0}_of_{1}_was_changed_to_{2}."
+#~ msgstr "Die Bestellung {0} von {1} wurde auf den {2} rückdatiert."
+
+#~ msgid "Change_order_status"
+#~ msgstr "Bestellstatus ändern"
+
 #~ msgid "Order_count"
 #~ msgstr "Anzahl Bestellungen"
 
@@ -1769,9 +1775,6 @@ msgstr "Das tatsächlich gelieferte Gewicht wird evtl. noch angepasst, d. h. der
 
 #~ msgid "Order_number_abbr"
 #~ msgstr "Bestell-Nr."
-
-#~ msgid "Not_picked_up"
-#~ msgstr "Nicht abgeholt"
 
 #~ msgid "Picked_up?"
 #~ msgstr "Abgeholt?"
