+++ plugins/Admin/webroot/js/admin.js
@@ -331,135 +331,6 @@ foodcoopshop.Admin = {
 
     },
 
-    createPickupDayCommentEditDialog: function (container) {
-
-        var dialogId = 'order-comment-edit-form';
-        var dialogHtml = foodcoopshop.DialogOrder.getHtmlForOrderCommentEdit(dialogId);
-        $(container).append(dialogHtml);
-
-        var buttons = {};
-        buttons['cancel'] = foodcoopshop.Helper.getJqueryUiCancelButton();
-        buttons['save'] = {
-            text: foodcoopshop.LocalizedJs.helper.save,
-            click: function() {
-                $('#order-comment-edit-form .ajax-loader').show();
-                $('.ui-dialog button').attr('disabled', 'disabled');
-
-                foodcoopshop.Helper.ajaxCall(
-                    '/admin/orders/editComment/',
-                    {
-                        orderId: $('#dialogOrderId').val(),
-                        orderComment: CKEDITOR.instances['dialogOrderComment'].getData()
-                    },
-                    {
-                        onOk: function (data) {
-                            document.location.reload();
-                        },
-                        onError: function (data) {
-                            console.log(data);
-                        }
-                    }
-                );
-            }
-        };
-
-        var dialog = $('#' + dialogId).dialog({
-            autoOpen: false,
-            height: 460,
-            width: 350,
-            modal: true,
-            close: function () {
-                $('#cke_dialogOrderComment').val('');
-                $('#dialogOrderId').val('');
-            },
-            buttons: buttons
-        });
-
-        return dialog;
-    },
-
-    initPickupDayCommentEditDialog: function (container) {
-
-        $('.pickup-day-comment-edit-button').on('click', function () {
-
-            foodcoopshop.Helper.destroyCkeditor('dialogOrderComment');
-            $('#pickup-day-comment-edit-form').remove();
-
-            var dialog = foodcoopshop.Admin.createPickupDayCommentEditDialog(container);
-            foodcoopshop.Helper.initCkeditor('dialogOrderComment');
-
-            var text = $(this).attr('originalTitle');
-            if (text == foodcoopshop.LocalizedJs.admin.AddComment) {
-                text = '';
-            }
-            CKEDITOR.instances['dialogOrderComment'].setData(text);
-            $('#pickup-day-comment-edit-form #dialogOrderId').val($(this).closest('tr').find('td:nth-child(1)').html());
-            dialog.dialog('open');
-
-        });
-
-    },
-
-    initOrderEditDialog: function (container) {
-
-        var dialogId = 'order-edit-form';
-        var dialogHtml = foodcoopshop.DialogOrder.getHtmlForOrderEdit(dialogId);
-        $(container).append(dialogHtml);
-
-        var buttons = {};
-        buttons['cancel'] = foodcoopshop.Helper.getJqueryUiCancelButton();
-        buttons['yes'] = {
-            text: foodcoopshop.LocalizedJs.helper.yes,
-            click: function() {
-                var newDate = $('#' + dialogId + ' .date-dropdown-placeholder select').val();
-
-                if (newDate == '' || $('#orderId').val() == '') {
-                    return false;
-                }
-
-                $('#order-edit-form .ajax-loader').show();
-                $('.ui-dialog button').attr('disabled', 'disabled');
-
-                foodcoopshop.Helper.ajaxCall(
-                    '/admin/orders/editDate/',
-                    {
-                        orderId: $('#orderId').val(),
-                        date: newDate
-                    },
-                    {
-                        onOk: function (data) {
-                            document.location.reload();
-                        },
-                        onError: function (data) {
-                            console.log(data);
-                        }
-                    }
-                );
-            }
-        };
-
-        var dialog = $('#' + dialogId).dialog({
-
-            autoOpen: false,
-            height: 280,
-            width: 350,
-            modal: true,
-
-            close: function () {
-                $('#' + dialogId + ' .date-dropdown-placeholder').html('');
-                $('#orderId').val('');
-            },
-            buttons: buttons
-        });
-
-        $('.edit-button').on('click', function () {
-            $('#' + dialogId + ' .date-dropdown-placeholder').html($(this).parent().parent().parent().parent().find('td.date-icon div.last-n-days-dropdown').html());
-            $('#' + dialogId + ' #orderId').val($(this).parent().parent().parent().parent().find('td:nth-child(1)').html());
-            dialog.dialog('open');
-        });
-
-    },
-
     initProductDepositEditDialog: function (container) {
 
         var dialogId = 'product-deposit-edit-form';
@@ -1317,104 +1188,6 @@ foodcoopshop.Admin = {
                 });
         });
 
-    },
-    
-    changeProductsPickedUpCallbackYes : function() {
-        foodcoopshop.Admin.changeProductsPickedUpCallback(1);
-    },
-
-    changeProductsPickedUpCallbackNo : function() {
-        foodcoopshop.Admin.changeProductsPickedUpCallback(0);
-    },
-    
-    changeProductsPickedUpCallback : function(state) {
-        
-        var dialogId = 'change-products-picked-up-form';
-        var customerIds = $('#' + dialogId + ' #customerId').val().split(',');
-        var pickupDay = $('input[name="pickupDay[]"]').val(); // filter-dropdown!
-        
-        $('#' + dialogId + ' .ajax-loader').show();
-        $('.ui-dialog button').attr('disabled', 'disabled');
-
-        foodcoopshop.Helper.ajaxCall(
-            '/admin/order-details/changeProductsPickedUp',
-            {
-                state: state,
-                customerIds: customerIds,
-                pickupDay: pickupDay
-            },
-            {
-                onOk: function (data) {
-                    document.location.reload();
-                },
-                onError: function (data) {
-                    var form = $('#' + dialogId);
-                    form.find('.ajax-loader').hide();
-                    foodcoopshop.Admin.appendFlashMessageToDialog(form, data.msg);
-                }
-            }
-        );
-        
-    },
-    
-    initChangeProductsPickedUpDialog : function(title, customerIds, customerName) {
-        
-        var dialogId = 'change-products-picked-up-form';
-        var dialogHtml = foodcoopshop.DialogOrderDetail.getHtmlForOrderDetailProductsPickupDayEdit(dialogId, title);
-        $(container).append(dialogHtml);
-
-        var buttons = {};
-        buttons['no'] = {
-            text: foodcoopshop.LocalizedJs.helper.no,
-            click: foodcoopshop.Admin.changeProductsPickedUpCallbackNo
-        };
-        buttons['yes'] = {
-            text: foodcoopshop.LocalizedJs.helper.yes,
-            click: foodcoopshop.Admin.changeProductsPickedUpCallbackYes
-        };
-        
-        var dialog = $('#' + dialogId).dialog({
-            autoOpen: false,
-            height: 230,
-            width: 500,
-            modal: true,
-            close: function () {
-                $('#customerId').val('');
-            },
-            buttons: buttons
-        });
-        
-        $('#' + dialogId + ' #customerId').val(customerIds.join(','));
-        
-        var infoMessage = $('#' + dialogId + ' p').html('');
-        var dialogHtml = '<p>' + foodcoopshop.LocalizedJs.admin.ThisInformationServesThePickupTeamToSeeWhoWasAlreadyHere + '</p>';
-        if (customerName != '') {
-            dialogHtml += '<p>' + foodcoopshop.LocalizedJs.admin.Member + ': <b>' + customerName + '</b></p>';
-        }
-        infoMessage.html(dialogHtml);
-        
-        dialog.dialog('open');        
-    },
-    
-    initChangeProductsPickedUpForAllCustomers: function(container) {
-        $('.change-products-picked-up-all-customers-button').on('click', function () {
-            var customerIds = [];
-            $(container).find('table.list tr.data').each(function() {
-                customerIds.push($(this).find('td:nth-child(2)').html());
-            });
-            var customerName = '';
-            var title = foodcoopshop.LocalizedJs.admin.WereTheProductsOfAllMembersPickedUp;
-            foodcoopshop.Admin.initChangeProductsPickedUpDialog(title, customerIds, customerName);
-        });
-    },
-    
-    initChangeProductsPickedUpByCustomer: function(container) {
-        $('.change-products-picked-up-button').on('click', function () {
-            var customerIds = [$(this).closest('tr').find('td:nth-child(2)').html()];
-            var customerName = $(this).closest('tr').find('td:nth-child(3)').text();
-            var title = foodcoopshop.LocalizedJs.admin.WereTheProductsPickedUp;
-            foodcoopshop.Admin.initChangeProductsPickedUpDialog(title, customerIds, customerName);
-        });
     },
 
     initManualOrderListSend: function (container, weekday) {
