+++ src/Locale/default.pot
@@ -5,7 +5,7 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
-"POT-Creation-Date: 2018-08-06 16:33+0200\n"
+"POT-Creation-Date: 2018-08-07 09:25+0200\n"
 "PO-Revision-Date: YYYY-mm-DD HH:MM+ZZZZ\n"
 "Last-Translator: NAME <EMAIL@ADDRESS>\n"
 "Language-Team: LANGUAGE <EMAIL@ADDRESS>\n"
@@ -653,13 +653,13 @@ msgstr ""
 msgid "The_member_needs_to_sign_again."
 msgstr ""
 
+msgid "Member"
+msgstr ""
+
 msgid "Were_the_products_picked_up?"
 msgstr ""
 
 msgid "This_information_serves_the_pickup_team_to_see_who_was_already_here."
-msgstr ""
-
-msgid "Member"
 msgstr ""
 
 msgid "All_products_picked_up?"
@@ -668,13 +668,7 @@ msgstr ""
 msgid "Were_the_products_of_all_members_picked_up?"
 msgstr ""
 
-msgid "Change_comment_of_order"
-msgstr ""
-
-msgid "Set_date_of_order_back_to"
-msgstr ""
-
-msgid "Set_date_of_order_back"
+msgid "Change_pickup_day_comment"
 msgstr ""
 
 msgid "Change_member_comment"
