+++ src/Locale/en_US/default.po
@@ -4,8 +4,8 @@
 msgid ""
 msgstr ""
 "Project-Id-Version: PROJECT VERSION\n"
-"POT-Creation-Date: 2018-08-06 16:33+0200\n"
-"PO-Revision-Date: 2018-08-06 16:33+0200\n"
+"POT-Creation-Date: 2018-08-07 09:25+0200\n"
+"PO-Revision-Date: 2018-08-07 09:22+0200\n"
 "Last-Translator: \n"
 "Language-Team: LANGUAGE <EMAIL@ADDRESS>\n"
 "Language: en_US\n"
@@ -654,14 +654,14 @@ msgstr "Change group for"
 msgid "The_member_needs_to_sign_again."
 msgstr "The member needs to sign in again."
 
+msgid "Member"
+msgstr "Member"
+
 msgid "Were_the_products_picked_up?"
 msgstr "Were the products picked up?"
 
 msgid "This_information_serves_the_pickup_team_to_see_who_was_already_here."
 msgstr "This input is helpful for the pickup team, but not mandatory."
-
-msgid "Member"
-msgstr "Member"
 
 msgid "All_products_picked_up?"
 msgstr "All products picked up?"
@@ -669,14 +669,9 @@ msgstr "All products picked up?"
 msgid "Were_the_products_of_all_members_picked_up?"
 msgstr "Were the products of all members picked up?"
 
-msgid "Change_comment_of_order"
-msgstr "Change order comment"
-
-msgid "Set_date_of_order_back_to"
-msgstr "Set date of order back to"
-
-msgid "Set_date_of_order_back"
-msgstr "Set date of order back"
+#, fuzzy
+msgid "Change_pickup_day_comment"
+msgstr "Change pickup day comment"
 
 msgid "Change_member_comment"
 msgstr "Change member comment"
@@ -1907,6 +1902,12 @@ msgstr "information-about-right-of-withdrawal"
 
 msgid "route_product"
 msgstr "product"
+
+#~ msgid "Set_date_of_order_back_to"
+#~ msgstr "Set date of order back to"
+
+#~ msgid "Set_date_of_order_back"
+#~ msgstr "Set date of order back"
 
 #~ msgid "Delivery_day"
 #~ msgstr "Delivery day"
